+++ src/config.ts
@@ -1,14 +1,17 @@
 import { SignOptions } from '@electron/windows-sign';
 
-export type CodesignOptions = Omit<SignOptions, 'appDirectory'>
+export type CodesignOptions = Omit<SignOptions, 'appDirectory'>;
 
 export interface MakerNSISConfig {
-  codesign?: CodesignOptions,
+  codesign?: CodesignOptions;
   updater?: {
-    url: string,
-    channel?: string,
-    updaterCacheDirName?: string,
-    publisherName?: string
-  }
-  getAppBuilderConfig?: () => Promise<Record<string, any>>
+    provider?: 'generic' | 'github';
+    url?: string;
+    owner?: string;
+    repo?: string;
+    channel?: string;
+    updaterCacheDirName?: string;
+    publisherName?: string | string[];
+  };
+  getAppBuilderConfig?: () => Promise<Record<string, any>>;
 }